Tidy routes
- move non-routes into _helpers.py
- move stats route into packager.py so the two routes have a shared namespace
